Extract panes
